extra

+++ Day_010/Day_010.py
@@ -36,14 +36,14 @@ operations={
 def calculator():
     print(logo)
 
-    num1=int(input("What is the first number?: "))
+    num1=float(input("What is the first number?: "))
     for key in operations:
         print(key)
     should_continue=True
 
     while should_continue:
         operation_symbol=input("Pick an operation from the line above: ")
-        num2=int(input("What is the next number?: "))
+        num2=float(input("What is the next number?: "))
         calculation_function=operations[operation_symbol]
         answer=calculation_function(num1,num2)
